fix(tech-debt): Split LogBrowser into focused mixins (Issue #92)
Split the 24-method LogBrowser class into focused mixins for better
separation of concerns:

- LogBrowserDisplayMixin (5 methods): Display formatting, status updates
- LogBrowserFilteringMixin (3 methods): Log content filtering, noise reduction
- LogBrowserNavigationMixin (6 methods): Cursor navigation, selection mode
- LogBrowser (10 methods): Core widget functionality
